requirements.txt
django
gunicorn
django-heroku
django-crispy-forms==1.8.1
dj-database-url==0.5.0